+++ modules/component-store/src/component-store.ts
@@ -34,9 +34,10 @@ import {
   isDevMode,
   Signal,
   computed,
+  type ValueEqualityFn,
 } from '@angular/core';
 import { isOnStateInitDefined, isOnStoreInitDefined } from './lifecycle_hooks';
-import { toSignal } from './to-signal';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 export interface SelectConfig {
   debounce?: boolean;
@@ -64,6 +65,13 @@ type SignalsProjector<Signals extends Signal<unknown>[], Result> = (
   }
 ) => Result;
 
+interface SelectSignalOptions<T> {
+  /**
+   * A comparison function which defines equality for select results.
+   */
+  equal?: ValueEqualityFn<T>;
+}
+
 @Injectable()
 export class ComponentStore<T extends object> implements OnDestroy {
   // Should be used only in ngOnDestroy.
@@ -79,7 +87,7 @@ export class ComponentStore<T extends object> implements OnDestroy {
   readonly state: Signal<T> = toSignal(
     this.state$.pipe(takeUntil(this.destroy$)),
     { requireSync: false, manualCleanup: true }
-  );
+  ) as Signal<T>;
   private ɵhasProvider = false;
 
   constructor(@Optional() @Inject(INITIAL_STATE_TOKEN) defaultState?: T) {
@@ -301,22 +309,37 @@ export class ComponentStore<T extends object> implements OnDestroy {
   /**
    * Creates a signal from the provided state projector function.
    */
-  selectSignal<Result>(projector: (state: T) => Result): Signal<Result>;
+  selectSignal<Result>(
+    projector: (state: T) => Result,
+    options?: SelectSignalOptions<Result>
+  ): Signal<Result>;
   /**
    * Creates a signal by combining provided signals.
    */
   selectSignal<Signals extends Signal<unknown>[], Result>(
-    ...signalsWithProjector: [
-      ...selectors: Signals,
-      projector: SignalsProjector<Signals, Result>
+    ...args: [...signals: Signals, projector: SignalsProjector<Signals, Result>]
+  ): Signal<Result>;
+  /**
+   * Creates a signal by combining provided signals.
+   */
+  selectSignal<Signals extends Signal<unknown>[], Result>(
+    ...args: [
+      ...signals: Signals,
+      projector: SignalsProjector<Signals, Result>,
+      options: SelectSignalOptions<Result>
     ]
   ): Signal<Result>;
   selectSignal(
     ...args:
-      | [(state: T) => unknown]
+      | [(state: T) => unknown, SelectSignalOptions<unknown>?]
       | [
           ...signals: Signal<unknown>[],
           projector: (...values: unknown[]) => unknown
+        ]
+      | [
+          ...signals: Signal<unknown>[],
+          projector: (...values: unknown[]) => unknown,
+          options: SelectSignalOptions<unknown>
         ]
   ): Signal<unknown> {
     if (args.length === 1) {
@@ -324,15 +347,31 @@ export class ComponentStore<T extends object> implements OnDestroy {
       return computed(() => projector(this.state()));
     }
 
-    const signals = args.slice(0, -1) as Signal<unknown>[];
-    const projector = args[args.length - 1] as (
+    const optionsOrProjector = args[args.length - 1] as (
+      ...values: unknown[]
+    ) => unknown | SelectSignalOptions<unknown>;
+    if (typeof optionsOrProjector === 'function') {
+      const signals = args.slice(0, -1) as Signal<unknown>[];
+
+      return computed(() => {
+        const values = signals.map((signal) => signal());
+        return optionsOrProjector(...values);
+      });
+    }
+
+    if (args.length === 2) {
+      const projector = args[0] as (state: T) => unknown;
+      return computed(() => projector(this.state()), optionsOrProjector);
+    }
+
+    const signals = args.slice(0, -2) as Signal<unknown>[];
+    const projector = args[args.length - 2] as (
       ...values: unknown[]
     ) => unknown;
-
     return computed(() => {
       const values = signals.map((signal) => signal());
       return projector(...values);
-    });
+    }, optionsOrProjector);
   }
 
   /**