+++ modules/data/schematics-core/utility/libs-version.ts
@@ -1,1 +1,1 @@
-export const platformVersion = '^16.0.0-beta.0';
+export const platformVersion = '^16.0.0-rc.0';