+++ modules/store/src/store.ts
@@ -2,12 +2,17 @@
 import { computed, Injectable, Provider, Signal } from '@angular/core';
 import { Observable, Observer, Operator } from 'rxjs';
 import { distinctUntilChanged, map, pluck } from 'rxjs/operators';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 import { ActionsSubject } from './actions_subject';
-import { Action, ActionReducer, FunctionIsNotAllowed } from './models';
+import {
+  Action,
+  ActionReducer,
+  SelectSignalOptions,
+  FunctionIsNotAllowed,
+} from './models';
 import { ReducerManager } from './reducer_manager';
 import { StateObservable } from './state';
-import { toSignal } from './to_signal';
 
 @Injectable()
 export class Store<T = object>
@@ -24,7 +29,7 @@ export class Store<T = object>
     super();
 
     this.source = state$;
-    this.state = toSignal(state$);
+    this.state = toSignal(state$, { manualCleanup: true });
   }
 
   select<K>(mapFn: (state: T) => K): Observable<K>;
@@ -101,9 +106,13 @@ export class Store<T = object>
    * Returns a signal of the provided selector.
    *
    * @param selector selector function
+   * @param options select signal options
    */
-  selectSignal<K>(selector: (state: T) => K): Signal<K> {
-    return computed(() => selector(this.state()));
+  selectSignal<K>(
+    selector: (state: T) => K,
+    options?: SelectSignalOptions<K>
+  ): Signal<K> {
+    return computed(() => selector(this.state()), { equal: options?.equal });
   }
 
   override lift<R>(operator: Operator<T, R>): Store<R> {